Refactored database class
