--- backend/app/models/leaderboard.py
@@ -1,11 +1,0 @@
-"""
-Models for leaderboard data structures
-"""
-
-from .punishment import PunishmentOut
-from .user import User
-
-
-class LeaderboardUser(User):
-    punishments: list[PunishmentOut]
-    total_value: int
