+++ backend/pylintrc
@@ -71,6 +71,8 @@ disable=too-many-lines,
         too-many-statements,
         too-many-arguments,
         too-many-public-methods,
+        too-many-instance-attributes,
+        missing-module-docstring,
         super-init-not-called,
         unnecessary-pass,
         invalid-name,
